# Performance Monitor: Guard Frame Rate