## Summary

I noticed that when fast-scrolling sometimes the UI frame rate would over-shoot significantly up to 200 FPS. This can happen when `const measuredRangeDuration = timestamp - droppedTimestamp;` drops below 100ms.  On a 120HZ screen frames are 8ms ~

For example if it calculated frames 100ms apart would lead to 100/20 (smoothingFrames) = 1000 / 5 => 200 FPS
The maximum should be 120 FPS / 60 FPS.

I wasn't sure if we should guard or remove the division by smoothingFrames.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to test out this change along with instructions how to run it and a description of the expected behavior. -->