@mobinni 2025-03-27
@piaskowyk Actually I wonder if we should be dividing by the amount of frames supported by the device.
120 HZ = 8 frames per second? Although this is actually determined by the display itself so sometimes it could run at 16. Maybe `smoothingFrames` should be 16 and not 20? 

```
const measuredRangeDuration = timestamp - droppedTimestamp;
const avgFrameTime = measuredRangeDuration / 16;
```

If your `measuredRangeDuration` is for example 100 MS => 100 / 16 = 6.2
1000 / 12.5 = 80 FPS on a 120hz display 🤔